- add toBigDecimal Method for roundoff in MethodUtils.

+++ app/src/main/java/com/android/pos/utils/MethodUtils.kt
@@ -600,11 +600,13 @@ class MethodUtils {
                 var tmp = value.toString()
                 var tmpIndex = tmp.indexOf(".", 0, true)
                 if (tmp.length > tmpIndex + 3) {
-                    return String.format("%.2f", value).toDouble()
+
+
+                    return String.format("%.2f", value).toBigDecimal().toDouble()
 
                 } else {
 
-                    return String.format("%.2f", value).toDouble()
+                    return String.format("%.2f", value).toBigDecimal().toDouble()
 
                 }
             } catch (e: java.lang.Exception) {
